Move command execution logic to the Runtime Client
